Task5. We must accept finite disappointment, but never lose infinite hope

+++ 0x0C-more_malloc_free/101-mul.c
@@ -10,7 +10,7 @@
  */
 int main(int argc, char *argv[])
 {
-	unsigned int h;
+	unsigned long h;
 	int x, y;
 
 	if (argc != 3)
